requirements.txt
Flask==1.1.2

# Testing and static analysis.
pytest==5.4.1
pytest-cov==2.8.1
flake8==3.7.9

# Data and workers.
redis==3.4.1
celery==4.4.2

# Forms.
Flask-WTF==0.14.3
WTForms-Alchemy==0.16.9
email-validator==1.0.5

# Extensions.
Flask-DebugToolbar==0.11.0
Flask-Mail==0.9.1
Flask-Login==0.5.0

# Environment variables
python-dotenv==0.13.0

# Production server
gunicorn==20.0.4

# Database
psycopg2==2.8.5
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.3